ASRgenomics-package {ASRgenomics}	R Documentation

ASRgenomics: Complementary Genomic Functions

Description

Presents a series of molecular and genetic routines in the R environment with the aim of assisting in analytical pipelines before and after the use of 'asreml' or another library to perform analyses such as Genomic Selection or Genome-Wide Association Analyses. Methods and examples are described in Gezan, Oliveira, Galli, and Murray (2022) https://asreml.kb.vsni.co.uk/wp-content/uploads/sites/3/ASRgenomics_Manual.pdf.

Author(s)

Maintainer: Salvador Gezan salvador.gezan@vsni.co.uk

Authors:

Darren Murray darren.murray@vsni.co.uk

Amanda Avelar de Oliveira support@vsni.co.uk

Giovanni Galli giovanni.galli@vsni.co.uk

Other contributors:

VSN International [copyright holder]

[Package ASRgenomics version 1.1.5 Index]