G.inverse {ASRgenomics}	R Documentation

Obtains the inverse of the genomic relationship matrix G

Description

Generates the inverse of a genomic relationship matrix \boldsymbol{G} that is provided. This input matrix should be of the full form (n \times n) with individual names assigned to rownames and colnames. Several checks for the stability of the matrix are presented based on the reciprocal conditional number.

In case of an ill-conditioned matrix, options of blending, bending or aligning before inverting are available. These options will be deprecated (discontinued) in future versions of ASRgenomics as they can be better implemented in the function G.tuneup().

Based on procedures published by Nazarian and Gezan et al. (2016).

Usage

G.inverse(
  G = NULL,
  A = NULL,
  rcn.thr = 0.000000000001,
  blend = FALSE,
  pblend = 0.02,
  bend = FALSE,
  eig.tol = NULL,
  align = FALSE,
  digits = 8,
  sparseform = FALSE,
  message = TRUE
)

Arguments

G
Input of the symmetric genomic relationship matrix \boldsymbol{G} in full form (n \times n), to obtain its inverse (default = NULL).

A
Input of the pedigree relationship matrix \boldsymbol{A} to perform blending or aligning in full form. It should be of the same dimension as the \boldsymbol{G} matrix (n \times n) (default = NULL) (to be deprecated).

rcn.thr
A threshold for identifying the \boldsymbol{G} matrix as an ill-conditioned matrix. Based on the reciprocal conditional number (default = 1e-12).

blend
If TRUE a "blending" with identity matrix \boldsymbol{I} or pedigree relationship matrix \boldsymbol{A} (if provided) is performed (default = FALSE) (to be deprecated).

pblend
If blending is requested this is the proportion of the identity matrix \boldsymbol{I} or pedigree relationship matrix \boldsymbol{A} to blend for (default = 0.02) (to be deprecated).

bend
If TRUE a "bending" is performed by making the matrix near positive definite (default = FALSE) (to be deprecated).

eig.tol
Defines relative positiveness (i.e., non-zero) of eigenvalues compared to the largest one. It determines which threshold of eigenvalues will be treated as zero (default = NULL) (to be deprecated).

align
If TRUE the genomic relationship matrix \boldsymbol{G} is aligned to the pedigree relationship matrix \boldsymbol{A} (default = FALSE) (to be deprecated).

digits
Set up the number of digits in used to round the output matrix (default = 8).

sparseform
If TRUE it generates an inverse matrix in sparse form to be used directly in asreml with required attributes (default = FALSE).

message
If TRUE diagnostic messages are printed on screen (default = TRUE).

Value

A list with three of the following elements:

Ginv: the inverse of \boldsymbol{G} matrix in full form (only if sparseform = FALSE).

Ginv.sparse: the inverse of \boldsymbol{G} matrix in sparse form (only if sparseform = TRUE).

status: the status (ill-conditioned or well-conditioned) of the inverse of \boldsymbol{G} matrix.

rcn: the reciprocal conditional number of the inverse of \boldsymbol{G} matrix.

References

Nazarian A., Gezan S.A. 2016. GenoMatrix: A software package for pedigree-based and genomic prediction analyses on complex traits. Journal of Heredity 107:372-379.

Examples

# Example: An ill-conditioned matrix.

# Get G matrix.
G <- G.matrix(M = geno.apple, method = "VanRaden")$G
G[1:5, 1:5]

# Get the inverse of G.
GINV <- G.inverse(G = G, bend = FALSE, blend = FALSE, align = FALSE)
GINV$Ginv[1:5, 1:5]
GINV$status

[Package ASRgenomics version 1.1.5 Index]