dummy.map_ {ASRgenomics}	R Documentation

Creates a dummy map if not provided

Description

Creates a dummy map if not provided

Usage

dummy.map_(marker.id = NULL, message = TRUE)

Arguments

marker.id
vector with names of markers to compose dummy map.

message
logical value indicating whether diagnostic messages should be printed on screen (default = TRUE).

Value

Data frame with dummy map. A single chromosome/linkage group is created and marker distances are a sequence from one to the number of markers.

[Package ASRgenomics version 1.1.5 Index]