geno.apple {ASRgenomics}	R Documentation

Genotypic data for apple dataset

Description

Genotypic data on 247 apple clones (i.e., genotypes) with a total of 2,828 SNP markers (coded as 0, 1, 2 and there are no missing records). Dataset obtained from supplementary material in Kumar et al. (2015).

Usage

geno.apple

Format

matrix

References

Kumar S., Molloy C., Muñoz P., Daetwyler H., Chagné D., and Volz R. 2015. Genome-enabled estimates of additive and nonadditive genetic variances and prediction of apple phenotypes across environments. G3 Genes, Genomes, Genetics 5:2711-2718.

Examples

geno.apple[1:5, 1:5]

[Package ASRgenomics version 1.1.5 Index]