geno.salmon {ASRgenomics}	R Documentation

Genotypic data for Atlantic salmon dataset

Description

Genotypic data on 1,481 Atlantic salmon samples. A total of 17,156 SNP markers (coded as 0, 1, 2 and NA for missing) are included in this dataset. Dataset obtained from supplementary material in Robledo et al. (2018).

Usage

geno.salmon

Format

matrix

References

Robledo D., Matika O., Hamilton A., and Houston R.D. 2018. Genome-wide association and genomic selection for resistance to amoebic gill disease in Atlantic salmon. G3 Genes, Genomes, Genetics 8:1195-1203.

Examples

geno.salmon[1:5, 1:5]

[Package ASRgenomics version 1.1.5 Index]