ped.salmon {ASRgenomics}	R Documentation

Pedigree data for Atlantic salmon dataset

Description

Pedigree data of 1,481 Atlantic salmon samples. Missing parental information coded as 0. Dataset obtained from supplementary material in Robledo et al. (2018).

Usage

ped.salmon

Format

data.frame

References

Robledo D., Matika O., Hamilton A., and Houston R.D. 2018. Genome-wide association and genomic selection for resistance to amoebic gill disease in Atlantic salmon. G3 Genes, Genomes, Genetics 8:1195-1203.

Examples

ped.salmon |> head()

[Package ASRgenomics version 1.1.5 Index]