pheno.apple {ASRgenomics}	R Documentation

Phenotypic data for apple dataset

Description

Phenotypic data on 247 apple clones (i.e., genotypes) evaluated for several fruit quality traits at two New Zealand sites, Motueka (MOT) and Hawkes Bay (HB). Dataset obtained from supplementary material in Kumar et al. (2015).

Usage

pheno.apple

Format

data.frame

References

Kumar S., Molloy C., Muñoz P., Daetwyler H., Chagné D., and Volz R. 2015. Genome-enabled estimates of additive and nonadditive genetic variances and prediction of apple phenotypes across environments. G3 Genes, Genomes, Genetics 5:2711–2718.

Examples

pheno.apple |> head()

[Package ASRgenomics version 1.1.5 Index]