pheno.salmon {ASRgenomics}	R Documentation

Phenotypic data for Atlantic salmon dataset

Description

Phenotypic data on 1,481 Atlantic salmon individuals. All fish were phenotyped for mean gill score (mean of the left gill and right gill scores) and amoebic load (qPCR values using Neoparamoeba perurans specific primers, amplified from one of the gills). Dataset obtained from supplementary material in Robledo et al. (2018).

Usage

pheno.salmon

Format

data.frame

References

Robledo D., Matika O., Hamilton A., and Houston R.D. 2018. Genome-wide association and genomic selection for resistance to amoebic gill disease in Atlantic salmon. G3 Genes, Genomes, Genetics 8:1195-1203.

Examples

pheno.salmon |> head()

[Package ASRgenomics version 1.1.5 Index]